feat: add zIndex to theme instead of separate util
